refactor: split resource_allocator.py into resource/ subpackage
- Created resource/ subpackage with modular structure:
  - types.py: ResourceType and AllocationStrategy enums
  - quota.py: Per-job CPU/memory quota tracking
  - gpu_pool.py: GPU reservation and inference slot management
  - admission.py: Admission control and overflow queuing
  - metrics.py: Resource utilization metrics collection
  - allocator.py: Core ResourceAllocator implementation
  - __init__.py: Public interface re-exporting all components

- Kept resource_allocator.py as backwards-compatible wrapper
  re-exporting from services.resource for existing imports

Audit findings for batch_queuing.py overlap:
- batch_queuing.py handles job batching/grouping and priority scheduling
- resource_allocator.py handles node-level resource allocation
- No significant overlap: batch_queuing.py does NOT do admission control
  or capacity tracking — it relies on resource_allocator for that
- batch_queuing.py remains unchanged

Closes #1277
